Do some refactor for the Python UDF code to try to reduce duplicate code. (#9902)
