+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/execution/python/GpuMapInBatchExec.scala
@@ -13,11 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/*** spark-rapids-shim-json-lines
-{"spark": "341db"}
-spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql.rapids.execution.python
+
+import scala.collection.JavaConverters.seqAsJavaListConverter
 
 import ai.rapids.cudf
 import ai.rapids.cudf.Table
@@ -62,6 +60,8 @@ trait GpuMapInBatchExec extends ShimUnaryExecNode with GpuPythonExecBase {
     val pythonRunnerConf = ArrowUtilsShim.getPythonRunnerConfMap(conf)
     val isPythonOnGpuEnabled = GpuPythonHelper.isPythonOnGpuEnabled(conf)
     val localOutput = output
+    val localBatchSize = batchSize
+    val localEvalType = pythonEvalType
 
     // Start process
     child.executeColumnar().mapPartitionsInternal { inputIter =>
@@ -77,8 +77,7 @@ trait GpuMapInBatchExec extends ShimUnaryExecNode with GpuPythonExecBase {
       }
 
       val pyInputIterator = new RebatchingRoundoffIterator(inputIter, pyInputTypes,
-          batchSize, numInputRows, numInputBatches)
-        .map { batch =>
+        localBatchSize, numInputRows, numInputBatches).map { batch =>
           // Here we wrap it via another column so that Python sides understand it
           // as a DataFrame.
           withResource(batch) { b =>
@@ -88,15 +87,16 @@ trait GpuMapInBatchExec extends ShimUnaryExecNode with GpuPythonExecBase {
               new ColumnarBatch(Array(gpuColumn), b.numRows())
             }
           }
-      }
+        }
       val pyRunner = new GpuArrowPythonRunner(
           chainedFunc,
-          pythonEvalType,
+          localEvalType,
           argOffsets,
           pyInputSchema,
           sessionLocalTimeZone,
           pythonRunnerConf,
-          batchSize) {
+          localBatchSize,
+          GpuColumnVector.structFromAttributes(localOutput.asJava)) {
         override def toBatch(table: Table): ColumnarBatch = {
           BatchGroupedIterator.extractChildren(table, localOutput)
         }
