+++ sql-plugin/src/main/spark341db/scala/org/apache/spark/sql/rapids/execution/python/shims/GpuCoGroupedArrowPythonRunner.scala
@@ -21,14 +21,13 @@ package org.apache.spark.sql.rapids.execution.python.shims
 
 import java.io.DataOutputStream
 
-import ai.rapids.cudf.{ArrowIPCWriterOptions, NvtxColor, NvtxRange, Table}
-import com.nvidia.spark.rapids.{GpuColumnVector, GpuSemaphore}
 import com.nvidia.spark.rapids.Arm.withResource
+import com.nvidia.spark.rapids.GpuSemaphore
 
 import org.apache.spark.{SparkEnv, TaskContext}
 import org.apache.spark.api.python.{ChainedPythonFunctions, PythonRDD, PythonWorker}
 import org.apache.spark.sql.execution.python.PythonUDFRunner
-import org.apache.spark.sql.rapids.execution.python._
+import org.apache.spark.sql.rapids.execution.python.{GpuArrowWriter, GpuPythonRunnerCommon}
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.vectorized.ColumnarBatch
 
@@ -47,34 +46,33 @@ class GpuCoGroupedArrowPythonRunner(
     timeZoneId: String,
     conf: Map[String, String],
     batchSize: Int,
-    pythonOutSchema: StructType)
-  extends GpuPythonRunnerBase[(ColumnarBatch, ColumnarBatch)](funcs, evalType, argOffsets)
-    with GpuPythonArrowOutput {
+    override val pythonOutSchema: StructType,
+    jobArtifactUUID: Option[String] = None)
+  extends GpuBasePythonRunner[(ColumnarBatch, ColumnarBatch)](funcs, evalType,
+    argOffsets, jobArtifactUUID) with GpuArrowPythonOutput with GpuPythonRunnerCommon {
 
   protected override def newWriter(
       env: SparkEnv,
-      worker: PythonWorker,
+      worker: PythonWorker,  // Changed from "Socket" to this "PythonWorker" from db341
       inputIterator: Iterator[(ColumnarBatch, ColumnarBatch)],
       partitionIndex: Int,
       context: TaskContext): Writer = {
     new Writer(env, worker, inputIterator, partitionIndex, context) {
 
       protected override def writeCommand(dataOut: DataOutputStream): Unit = {
-
         // Write config for the worker as a number of key -> value pairs of strings
         dataOut.writeInt(conf.size)
         for ((k, v) <- conf) {
           PythonRDD.writeUTF(k, dataOut)
           PythonRDD.writeUTF(v, dataOut)
         }
-
         PythonUDFRunner.writeUDFs(dataOut, funcs, argOffsets)
       }
 
       override def writeNextInputToStream(dataOut: DataOutputStream): Boolean = {
-        // For each we first send the number of dataframes in each group then send
-        // first df, then send second df.  End of data is marked by sending 0.
         if (inputIterator.hasNext) {
+          // For each we first send the number of dataframes in each group then send
+          // first df, then send second df.
           dataOut.writeInt(2)
           val (leftGroupBatch, rightGroupBatch) = inputIterator.next()
           withResource(Seq(leftGroupBatch, rightGroupBatch)) { _ =>
@@ -85,6 +83,7 @@ class GpuCoGroupedArrowPythonRunner(
         } else {
           // The iterator can grab the semaphore even on an empty batch
           GpuSemaphore.releaseIfNecessary(TaskContext.get())
+          //  End of data is marked by sending 0.
           dataOut.writeInt(0)
           false
         }
@@ -92,42 +91,20 @@ class GpuCoGroupedArrowPythonRunner(
 
       private def writeGroupBatch(groupBatch: ColumnarBatch, batchSchema: StructType,
           dataOut: DataOutputStream): Unit = {
-        val writer = {
-          val builder = ArrowIPCWriterOptions.builder()
-          builder.withMaxChunkSize(batchSize)
-          builder.withCallback((table: Table) => {
-            table.close()
-            GpuSemaphore.releaseIfNecessary(TaskContext.get())
-          })
-          // Flatten the names of nested struct columns, required by cudf arrow IPC writer.
-          GpuPythonRunnerUtils.flattenNames(batchSchema).foreach { case (name, nullable) =>
-            if (nullable) {
-              builder.withColumnNames(name)
-            } else {
-              builder.withNotNullableColumnNames(name)
-            }
-          }
-          Table.writeArrowIPCChunked(builder.build(), new BufferToStreamWriter(dataOut))
-        }
+        val gpuArrowWriter = GpuArrowWriter(batchSchema, batchSize)
         try {
-          withResource(new NvtxRange("write python batch", NvtxColor.DARK_GREEN)) { _ =>
-            // The callback will handle closing table and releasing the semaphore
-            writer.write(GpuColumnVector.from(groupBatch))
-          }
+          gpuArrowWriter.start(dataOut)
+          gpuArrowWriter.write(groupBatch)
         } catch {
           case t: Throwable =>
             // release the semaphore in case of exception in the middle of writing a batch
             GpuSemaphore.releaseIfNecessary(TaskContext.get())
             throw t
         } finally {
-          writer.close()
+          gpuArrowWriter.reset()
           dataOut.flush()
         }
-      } // end of writeGroup
+      } // end of writeGroupBatch
     }
-  } // end of newWriterThread
-
-  def toBatch(table: Table): ColumnarBatch = {
-    GpuColumnVector.from(table, GpuColumnVector.extractTypes(pythonOutSchema))
   }
 }
