+++ sql-plugin/src/main/spark341db/scala/org/apache/spark/sql/rapids/execution/python/shims/GpuGroupUDFArrowPythonRunner.scala
@@ -24,15 +24,13 @@ package org.apache.spark.sql.rapids.execution.python.shims
 
 import java.io.DataOutputStream
 
-import ai.rapids.cudf._
-import com.nvidia.spark.rapids._
-import com.nvidia.spark.rapids.Arm.withResource
+import com.nvidia.spark.rapids.GpuSemaphore
 
 import org.apache.spark.{SparkEnv, TaskContext}
 import org.apache.spark.api.python._
 import org.apache.spark.sql.execution.python.PythonUDFRunner
-import org.apache.spark.sql.rapids.execution.python._
-import org.apache.spark.sql.types._
+import org.apache.spark.sql.rapids.execution.python.{GpuArrowPythonWriter, GpuPythonRunnerCommon}
+import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.vectorized.ColumnarBatch
 
 /**
@@ -56,61 +54,36 @@ class GpuGroupUDFArrowPythonRunner(
     timeZoneId: String,
     conf: Map[String, String],
     batchSize: Long,
-    pythonOutSchema: StructType)
-  extends GpuPythonRunnerBase[ColumnarBatch](funcs, evalType, argOffsets)
-    with GpuPythonArrowOutput {
+    override val pythonOutSchema: StructType,
+    jobArtifactUUID: Option[String] = None)
+  extends GpuBasePythonRunner[ColumnarBatch](funcs, evalType, argOffsets, jobArtifactUUID)
+    with GpuArrowPythonOutput with GpuPythonRunnerCommon {
 
   protected override def newWriter(
       env: SparkEnv,
-      worker: PythonWorker,
+      worker: PythonWorker, // From DB341, changed from Socket to PythonWorker
       inputIterator: Iterator[ColumnarBatch],
       partitionIndex: Int,
       context: TaskContext): Writer = {
     new Writer(env, worker, inputIterator, partitionIndex, context) {
 
-      protected override def writeCommand(dataOut: DataOutputStream): Unit = {
-
-        // Write config for the worker as a number of key -> value pairs of strings
-        dataOut.writeInt(conf.size)
-        for ((k, v) <- conf) {
-          PythonRDD.writeUTF(k, dataOut)
-          PythonRDD.writeUTF(v, dataOut)
+      val arrowWriter = new GpuArrowPythonWriter(pythonInSchema, batchSize) {
+        override protected def writeUDFs(dataOut: DataOutputStream): Unit = {
+          PythonUDFRunner.writeUDFs(dataOut, funcs, argOffsets)
         }
+      }
 
-        PythonUDFRunner.writeUDFs(dataOut, funcs, argOffsets)
+      protected override def writeCommand(dataOut: DataOutputStream): Unit = {
+        arrowWriter.writeCommand(dataOut, conf)
       }
 
       override def writeNextInputToStream(dataOut: DataOutputStream): Boolean = {
-        // write out number of columns
         try {
           if (inputIterator.hasNext) {
-            val builder = ArrowIPCWriterOptions.builder()
-            builder.withMaxChunkSize(batchSize)
-            builder.withCallback((table: Table) => {
-              table.close()
-              GpuSemaphore.releaseIfNecessary(TaskContext.get())
-            })
-            // Flatten the names of nested struct columns, required by cudf Arrow IPC writer.
-            GpuPythonRunnerUtils.flattenNames(pythonInSchema).foreach { case (name, nullable) =>
-              if (nullable) {
-                builder.withColumnNames(name)
-              } else {
-                builder.withNotNullableColumnNames(name)
-              }
-            }
-            val writer = {
-              // write 1 out to indicate there is more to read
-              dataOut.writeInt(1)
-              Table.writeArrowIPCChunked(builder.build(), new BufferToStreamWriter(dataOut))
-            }
-            val table = withResource(inputIterator.next()) { nextBatch =>
-              GpuColumnVector.from(nextBatch)
-            }
-            withResource(new NvtxRange("write python batch", NvtxColor.DARK_GREEN)) { _ =>
-              // The callback will handle closing table and releasing the semaphore
-              writer.write(table)
-            }
-            writer.close()
+            dataOut.writeInt(1)
+            arrowWriter.start(dataOut)
+            arrowWriter.writeAndClose(inputIterator.next())
+            arrowWriter.reset()
             dataOut.flush()
             true
           } else {
@@ -123,15 +96,12 @@ class GpuGroupUDFArrowPythonRunner(
           }
         } catch {
           case t: Throwable =>
+            arrowWriter.close()
             // release the semaphore in case of exception in the middle of writing a batch
             GpuSemaphore.releaseIfNecessary(TaskContext.get())
             throw t
         }
       }
     }
-  }
-
-  def toBatch(table: Table): ColumnarBatch = {
-    GpuColumnVector.from(table, GpuColumnVector.extractTypes(pythonOutSchema))
   }
 }
